+++ src/main/java/com/sydders/wellspring/datagen/ModBlockTagsProvider.java
@@ -116,6 +116,12 @@ public class ModBlockTagsProvider extends BlockTagsProvider {
                 .add(ModBlocks.POTTED_SIFT_SAPLING.get())
                 .add(ModBlocks.POTTED_WITHERED_SAPLING.get());
 
+        tag(BlockTags.FLOWER_POTS)
+                .add(ModBlocks.CROTON.get());
+
+        tag(BlockTags.SUPPORTS_DRY_VEGETATION)
+                .add(ModBlocks.SIFT_STONE.get());
+
         tag(ModTags.Blocks.NEEDS_BAZULIUM_TOOL)
                 .addTag(BlockTags.NEEDS_DIAMOND_TOOL);