+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Features
+
 ## Entities
 * Updated Blub model
 * Updated Blub animations
+
+## Blocks
+* Added croton plant